+++ Libft/srcs/strings/ft_longlen.c
@@ -6,13 +6,13 @@
 /*   By: vtestut <vtestut@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 17:44:45 by vtestut           #+#    #+#             */
-/*   Updated: 2022/12/05 17:45:03 by vtestut          ###   ########.fr       */
+/*   Updated: 2022/12/07 17:51:37 by vtestut          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_longlen(size_t n)
+int	ft_longlen(unsigned int n)
 {
 	int	len;
